style: structure the app
structure the app by grouping the related methods in a seperate file
