+++ .pre-commit-config.yaml
@@ -14,7 +14,7 @@ repos:
       - id: flake8
         args: ['--max-line-length=120', '--ignore=W']
 
-  - repo: https://github.com/pre-commit/mirrors-isort
-    rev: v4.3.21
+  - repo: https://github.com/pycqa/isort
+    rev: 5.12.0
     hooks:
       - id: isort
